Moved cellular_automata subpackage to landlab.ca.
