+++ landlab/ca/examples/isotropic_turbulent_suspension.py
@@ -1,19 +1,21 @@
 #!/usr/env/python
 
 """
-little_ca_test.py
+isotropic_turbulent_suspension.py
 
-Tests the creation and execution of a CellLab-CTS model, by creating a
-simple two-state CA on a small grid.
+Example of a continuous-time, stochastic, pair-based cellular automaton model,
+which simulates the diffusion of suspended, neutrally buoyant particles in a
+turbulent fluid.
 
-Created by Greg Tucker, May 2015
+Written by Greg Tucker, February 2015
 """
-from __future__ import print_function
 
 import time
+import matplotlib
+from numpy import where
 from landlab import RasterModelGrid
-from landlab.components.cellular_automata.celllab_cts import Transition, CAPlotter
-from landlab.components.cellular_automata.raster_cts import RasterCTS
+from landlab.ca.celllab_cts import Transition, CAPlotter
+from landlab.ca.raster_cts import RasterCTS
 
 
 def setup_transition_list():
@@ -22,6 +24,10 @@ def setup_transition_list():
     transitions for a biased random walk, in which the rate of downward
     motion is greater than the rate in the other three directions.
 
+    Parameters
+    ----------
+    (none)
+
     Returns
     -------
     xn_list : list of Transition objects
@@ -29,20 +35,17 @@ def setup_transition_list():
 
     Notes
     -----
-    This doesn't represent any particular process, but rather is simply used
-    to test the CA code. The transition rules have 0-0 pairs transitioning to
-    0-1 or 1-0 pairs (50/50 chance) and thence to 1-1 pairs, at which point
-    there are no further transitions.
+    State 0 represents fluid and state 1 represents a particle (such as a
+    sediment grain, tea leaf, or dissolved heavy particle).
 
     The states and transitions are as follows:
 
     Pair state      Transition to       Process             Rate (cells/s)
     ==========      =============       =======             ==============
-    0 (0-0)         1 (0-1)                                 0.5
-                    2 (1-0)                                 0.5
-    1 (0-1)         3 (1-1)                                 1.0
-    2 (1-0)         3 (1-1)                                 1.0
-    3 (1-1)         (none)                                  -
+    0 (0-0)         (none)              -                   -
+    1 (0-1)         2 (1-0)             left/down motion    10.0
+    2 (1-0)         1 (0-1)             right/up motion     10.0
+    3 (1-1)         (none)              -                   -
 
     """
 
@@ -57,10 +60,8 @@ def setup_transition_list():
     #    (left/bottom cell, right/top cell, orientation)
     #  - Transition rate (cells per time step, in this case 1 sec)
     #  - Name for transition
-    xn_list.append( Transition((0,0,0), (0,1,0), 0.5, '') )
-    xn_list.append( Transition((0,0,0), (1,0,0), 0.5, '') )
-    xn_list.append( Transition((0,1,0), (1,1,0), 1., '') )
-    xn_list.append( Transition((1,0,0), (1,1,0), 1., '') )
+    xn_list.append( Transition((0,1,0), (1,0,0), 10., 'left/down motion') )
+    xn_list.append( Transition((1,0,0), (0,1,0), 10., 'right/up motion') )
 
     return xn_list
 
@@ -70,10 +71,10 @@ def main():
     # INITIALIZE
 
     # User-defined parameters
-    nr = 5  # number of rows in grid
-    nc = 5  # number of columns in grid
-    plot_interval = 10.0   # time interval for plotting, sec
-    run_duration = 10.0   # duration of run, sec
+    nr = 80  # number of rows in grid
+    nc = 50  # number of columns in grid
+    plot_interval = 0.5   # time interval for plotting, sec
+    run_duration = 0.0   # duration of run, sec
     report_interval = 10.0  # report interval, in real-time seconds
 
     # Remember the clock time, and calculate when we next want to report
@@ -88,11 +89,16 @@ def main():
     mg.set_closed_boundaries_at_grid_edges(True, True, True, True)
 
     # Set up the states and pair transitions.
-    ns_dict = { 0 : 'black', 1 : 'white' }
+    ns_dict = { 0 : 'fluid', 1 : 'particle' }
     xn_list = setup_transition_list()
 
     # Create the node-state array and attach it to the grid
     node_state_grid = mg.add_zeros('node', 'node_state_map', dtype=int)
+
+    # Initialize the node-state array: here, the initial condition is a pile of
+    # resting grains at the bottom of a container.
+    bottom_rows = where(mg.node_y<0.1*nr)[0]
+    node_state_grid[bottom_rows] = 1
 
     # For visual display purposes, set all boundary nodes to fluid
     node_state_grid[mg.closed_boundary_nodes] = 0
@@ -100,8 +106,14 @@ def main():
     # Create the CA model
     ca = RasterCTS(mg, ns_dict, xn_list, node_state_grid)
 
+    # Set up colors for plotting
+    grain = '#5F594D'
+    fluid = '#D0E4F2'
+    clist = [fluid,grain]
+    my_cmap = matplotlib.colors.ListedColormap(clist)
+
     # Create a CAPlotter object for handling screen display
-    ca_plotter = CAPlotter(ca)
+    ca_plotter = CAPlotter(ca, cmap=my_cmap)
 
     # Plot the initial grid
     ca_plotter.update_plot()
@@ -114,13 +126,12 @@ def main():
         # know that the sim is running ok
         current_real_time = time.time()
         if current_real_time >= next_report:
-            print('Current sim time', current_time, '(',
-                  100 * current_time/run_duration, '%)')
+            print('Current simulation time '+str(current_time)+' ('+str(int(100*current_time/run_duration))+'%)')
             next_report = current_real_time + report_interval
 
         # Run the model forward in time until the next output step
         ca.run(current_time+plot_interval, ca.node_state,
-               plot_each_transition=True, plotter=ca_plotter)
+               plot_each_transition=False)
         current_time += plot_interval
 
         # Plot the current grid
@@ -132,12 +143,7 @@ def main():
     # Plot
     ca_plotter.finalize()
 
-    print('ok, here are the keys')
-    print(ca.__dict__.keys())
 
 # If user runs this file, activate the main() function
 if __name__ == "__main__":
-    #import cProfile
-    #fname = 'test_profiler_for_little_ca.txt'
-    #cProfile.run('print main(); print') #, fname)
     main()
