+++ landlab/ca/examples/settle_test.py
@@ -1,30 +1,25 @@
-#!/usr/env/python
-
 """
-diffusion_in_gravity.py
+settle_test.py
 
-Example of a continuous-time, stochastic, pair-based cellular automaton model,
-which simulates diffusion by random particle motion in a gravitational field.
-The purpose of the example is to demonstrate the use of an OrientedRasterLCA.
+Tests "external update" capability in landlab_ca.py.
 
-GT, September 2014
+GT Nov 2014
 """
 from __future__ import print_function
 
 _DEBUG = False
 
 import time
-from numpy import where, bitwise_and
+from numpy import where
 from landlab import RasterModelGrid
-from landlab.components.cellular_automata.celllab_cts import Transition, CAPlotter
-from landlab.components.cellular_automata.oriented_raster_cts import OrientedRasterCTS
-
+from landlab.ca.raster_cts import RasterCTS
+from landlab.ca.celllab_cts import Transition, CAPlotter
+from landlab.ca.oriented_raster_cts import OrientedRasterCTS
 
 def setup_transition_list():
     """
     Creates and returns a list of Transition() objects to represent state
-    transitions for a biased random walk, in which the rate of downward
-    motion is greater than the rate in the other three directions.
+    transitions for particles that simply settle under gravity.
 
     Parameters
     ----------
@@ -37,29 +32,22 @@ def setup_transition_list():
 
     Notes
     -----
-    State 0 represents fluid and state 1 represents a particle (such as a
-    sediment grain or dissolved heavy particle).
-
     The states and transitions are as follows:
 
-    Pair state      Transition to       Process         Rate
-    ==========      =============       =======         ====
-    0 (0-0)         (none)              -               -
-    1 (0-1)         2 (1-0)             left motion     1.0
-    2 (1-0)         1 (0-1)             right motion    1.0
-    3 (1-1)         (none)              -               -
-    4 (0/0)         (none)              -               -
-    5 (0/1)         6 (1/0)             down motion     1.1
-    6 (1/0)         5 (0/1)             up motion       0.9
-    7 (1/1)         (none)              -               -
+    Pair state      Transition to       Process     Rate
+    ==========      =============       =======     ====
+    0 (0-0,0)
+    1 (0-1,0)
+    2 (1-0,0)
+    3 (1-1,0)
+    4 (0-0,1)
+    5 (0-1,1)
+    1 (0-1)         2 (1-0)             settling    1.0
 
     """
     xn_list = []
 
-    xn_list.append( Transition((0,1,0), (1,0,0), 1., 'left motion') )
-    xn_list.append( Transition((1,0,0), (0,1,0), 1., 'right motion') )
-    xn_list.append( Transition((0,1,1), (1,0,1), 1.1, 'down motion') )
-    xn_list.append( Transition((1,0,1), (0,1,1), 0.9, 'up motion') )
+    xn_list.append( Transition(5, 6, 1., 'settling') )
 
     if _DEBUG:
         print()
@@ -75,10 +63,10 @@ def main():
     # INITIALIZE
 
     # User-defined parameters
-    nr = 80
-    nc = 80
-    plot_interval = 2
-    run_duration = 200
+    nr = 10
+    nc = 10
+    plot_interval = 0.25
+    run_duration = 40.0
     report_interval = 5.0  # report interval, in real-time seconds
 
     # Remember the clock time, and calculate when we next want to report
@@ -88,23 +76,24 @@ def main():
 
     # Create grid
     mg = RasterModelGrid(nr, nc, 1.0)
-
-    # Make the boundaries be walls
     mg.set_closed_boundaries_at_grid_edges(True, True, True, True)
 
     # Set up the states and pair transitions.
-    ns_dict = { 0 : 'fluid', 1 : 'particle' }
+    # Transition data here represent a body of fractured rock, with rock
+    # represented by nodes with state 0, and saprolite (weathered rock)
+    # represented by nodes with state 1. Node pairs (links) with 0-1 or 1-0
+    # can undergo a transition to 1-1, representing chemical weathering of the
+    # rock.
+    ns_dict = { 0 : 'air', 1 : 'particle' }
     xn_list = setup_transition_list()
 
     # Create the node-state array and attach it to the grid
     node_state_grid = mg.add_zeros('node', 'node_state_map', dtype=int)
+    node_state_grid[where(mg.node_y>nr-3)[0]] = 1
 
-    # Initialize the node-state array
-    middle_rows = where(bitwise_and(mg.node_y>0.45*nr, mg.node_y<0.55*nr))[0]
-    node_state_grid[middle_rows] = 1
-
-    # Create the CA model
+        # Create the CA model
     ca = OrientedRasterCTS(mg, ns_dict, xn_list, node_state_grid)
+    #ca = RasterCTS(mg, ns_dict, xn_list, node_state_grid)
 
     # Debug output if needed
     if _DEBUG:
@@ -123,6 +112,7 @@ def main():
 
     # RUN
     current_time = 0.0
+    updated = False
     while current_time < run_duration:
 
         # Once in a while, print out simulation and real time to let the user
@@ -136,6 +126,13 @@ def main():
         ca.run(current_time+plot_interval, ca.node_state,
                plot_each_transition=False) #, plotter=ca_plotter)
         current_time += plot_interval
+
+        # Add a bunch of particles
+        if current_time > run_duration/2. and not updated:
+            print('updating...')
+            node_state_grid[where(ca.grid.node_y>(nc/2.0))[0]] = 1
+            ca.update_link_states_and_transitions(current_time)
+            updated = True
 
         # Plot the current grid
         ca_plotter.update_plot()
