+++ landlab/ca/hex_lattice_tectonicizer.py
