+++ landlab/ca/tests/test_celllab_cts.py
@@ -10,11 +10,11 @@ Created on Thu Jul  9 08:20:06 2015
 from nose.tools import assert_equal
 from numpy.testing import assert_array_equal
 from landlab import RasterModelGrid, HexModelGrid
-from landlab.components.cellular_automata.celllab_cts import Transition, Event
-from landlab.components.cellular_automata.raster_cts import RasterCTS
-from landlab.components.cellular_automata.oriented_raster_cts import OrientedRasterCTS
-from landlab.components.cellular_automata.hex_cts import HexCTS
-from landlab.components.cellular_automata.oriented_hex_cts import OrientedHexCTS
+from landlab.ca.celllab_cts import Transition, Event
+from landlab.ca.raster_cts import RasterCTS
+from landlab.ca.oriented_raster_cts import OrientedRasterCTS
+from landlab.ca.hex_cts import HexCTS
+from landlab.ca.oriented_hex_cts import OrientedHexCTS
 from heapq import heappush
 from heapq import heappop
